Refactor DAGs visualizer (#1968)
* Move ws and branchID resolve function to utils directory

* Extract constants

* Move utxo data structures to models

* Move hashString from utxoStores to utils

* Implement cytoscapejs library

* Replace drawings in UTXOStores with wrapped cytoscapeLib

* Move branch data structures to models

* Replace drawings in BranchStores with wrapped cytoscapeLib

* Replace drawings in TangleStore with wrapped vivagraphLib

* Fix using wrong function call from vivaGraphLib

* Remove unused variables

* Use appropriate function from vivagraphLib

* Use appropriate function from vivagraphLib

* Extract LinkToDashboard a component

* Add missing unregisterHandlers in Branch and UTXO stores

* Unregister TransactionBooked handler

* Move list limit checking to a new function

* Fix vertices limiting issue in UTXO and Branch DAG

* Center searched vertex

* Wrap cytoscapejs event listener

* Move websocket connection to Root component

* Initialize MarkersManager events

* Update pkged.go

* Add full url to dashboard in global setting

* Check nodeUI existence first in vivagraph unselectVertex
