+++ plugins/dagsvisualizer/frontend/src/stores/BranchStore.tsx
@@ -1,51 +1,22 @@
 import { action, makeObservable, observable, ObservableMap } from 'mobx';
-import { registerHandler, unregisterHandler, WSMsgType } from '../WS';
-import cytoscape from 'cytoscape';
+import { registerHandler, unregisterHandler, WSMsgType } from 'utils/WS';
+import { MAX_VERTICES } from 'utils/constants';
 import dagre from 'cytoscape-dagre';
-import { dagreOptions } from 'styles/graphStyle';
 import layoutUtilities from 'cytoscape-layout-utilities';
-
-export class branchVertex {
-    ID: string;
-    type: string;
-    parents: Array<string>;
-    isConfirmed: boolean;
-    conflicts: conflictBranches;
-    gof: string;
-    aw: number;
-}
-
-export class conflictBranches {
-    branchID: string;
-    conflicts: Array<conflict>;
-}
-
-export class conflict {
-    outputID: any;
-    branchIDs: Array<string>;
-}
-export class branchParentUpdate {
-    ID: string;
-    parents: Array<string>;
-}
-
-export class branchConfirmed {
-    ID: string;
-}
-
-export class branchWeightChanged {
-    ID: string;
-    weight: number;
-    gof: string;
-}
+import { cytoscapeLib, drawBranch, initBranchDAG } from 'graph/cytoscape';
+import {
+    branchVertex,
+    branchParentUpdate,
+    branchConfirmed,
+    branchWeightChanged
+} from 'models/branch';
 
 export class BranchStore {
-    @observable maxBranchVertices = 500;
+    @observable maxBranchVertices = MAX_VERTICES;
     @observable branches = new ObservableMap<string, branchVertex>();
     @observable selectedBranch: branchVertex = null;
     @observable paused = false;
     @observable search = '';
-    @observable explorerAddress = 'localhost:8081';
     branchOrder: Array<any> = [];
     draw = true;
 
@@ -53,10 +24,8 @@ export class BranchStore {
     branchToRemoveAfterResume = [];
     branchToAddAfterResume = [];
 
-    cy;
-    layout;
-    layoutApi;
     layoutUpdateTimerID;
+    graph;
 
     constructor() {
         makeObservable(this);
@@ -67,18 +36,31 @@ export class BranchStore {
             WSMsgType.BranchWeightChanged,
             this.branchWeightChanged
         );
-
-        cytoscape.use(dagre);
-        cytoscape.use(layoutUtilities);
     }
 
     unregisterHandlers() {
         unregisterHandler(WSMsgType.Branch);
         unregisterHandler(WSMsgType.BranchParentsUpdate);
+        unregisterHandler(WSMsgType.BranchConfirmed);
+        unregisterHandler(WSMsgType.BranchWeightChanged);
     }
 
     @action
     addBranch = (branch: branchVertex) => {
+        this.checkLimit();
+
+        this.branchOrder.push(branch.ID);
+        this.branches.set(branch.ID, branch);
+
+        if (this.paused) {
+            this.branchToAddAfterResume.push(branch.ID);
+        }
+        if (this.draw) {
+            this.drawVertex(branch);
+        }
+    };
+
+    checkLimit = () => {
         if (this.branchOrder.length >= this.maxBranchVertices) {
             const removed = this.branchOrder.shift();
             this.branches.delete(removed);
@@ -89,16 +71,6 @@ export class BranchStore {
             } else {
                 this.removeVertex(removed);
             }
-        }
-
-        this.branchOrder.push(branch.ID);
-        this.branches.set(branch.ID, branch);
-
-        if (this.paused) {
-            this.branchToAddAfterResume.push(branch.ID);
-        }
-        if (this.draw) {
-            this.drawVertex(branch);
         }
     };
 
@@ -146,7 +118,7 @@ export class BranchStore {
     clearSelected = (removePreSelectedNode?: boolean) => {
         // unselect preselected node manually
         if (removePreSelectedNode && this.selectedBranch) {
-            this.cy.getElementById(this.selectedBranch.ID).unselect();
+            this.graph.unselectVertex(this.selectedBranch.ID);
         }
 
         this.selectedBranch = null;
@@ -165,6 +137,7 @@ export class BranchStore {
     @action
     updateVerticesLimit = (num: number) => {
         this.maxBranchVertices = num;
+        this.trimBranchToVerticesLimit();
     };
 
     @action
@@ -177,33 +150,11 @@ export class BranchStore {
         if (!this.search) return;
 
         this.selectBranch(this.search);
+        this.centerBranch(this.search);
     };
 
     getBranchVertex = (branchID: string) => {
         return this.branches.get(branchID);
-    };
-
-    selectBranch = (branchID: string) => {
-        // clear pre-selected branch.
-        this.clearSelected(true);
-
-        const branchNode = this.cy.getElementById(branchID);
-        if (!branchNode) return;
-        // select the node manually
-        branchNode.select();
-        this.centerBranch(branchID);
-
-        this.updateSelected(branchID);
-    };
-
-    centerBranch = (branchID: string) => {
-        const branchNode = this.cy.getElementById(branchID);
-        if (!branchNode) return;
-        this.cy.center(branchNode);
-    };
-
-    updateExplorerAddress = (addr: string) => {
-        this.explorerAddress = addr;
     };
 
     drawExistedBranches = () => {
@@ -214,14 +165,6 @@ export class BranchStore {
 
     updateDrawStatus = (draw: boolean) => {
         this.draw = draw;
-    };
-
-    clearGraph = () => {
-        this.cy.elements().remove();
-    };
-
-    centerEntireGraph = () => {
-        this.cy.center();
     };
 
     resumeAndSyncGraph = () => {
@@ -241,87 +184,67 @@ export class BranchStore {
         this.branchToRemoveAfterResume = [];
     };
 
-    removeVertex = (branchID: string) => {
-        this.vertexChanges++;
-        const uiID = '#' + branchID;
-        this.cy.remove(uiID);
-    };
-
     drawVertex = (branch: branchVertex) => {
         this.vertexChanges++;
 
-        const v = this.cy.add({
-            group: 'nodes',
-            data: { id: branch.ID }
-        });
+        drawBranch(branch, this.graph, this.branches);
+    };
 
-        if (branch.parents) {
-            branch.parents.forEach(pID => {
-                const b = this.branches.get(pID);
-                if (b) {
-                    this.cy.add({
-                        group: 'edges',
-                        data: { source: pID, target: branch.ID }
-                    });
-                }
-            });
-        }
+    removeVertex = (branchID: string) => {
+        this.vertexChanges++;
+        this.graph.removeVertex(branchID);
+    };
 
-        this.layoutApi.placeNewNodes(v);
+    selectBranch = (branchID: string) => {
+        // clear pre-selected branch.
+        this.clearSelected(true);
+        this.graph.selectVertex(branchID);
+        this.updateSelected(branchID);
+    };
+
+    centerBranch = (branchID: string) => {
+        this.graph.centerVertex(branchID);
+    };
+
+    centerEntireGraph = () => {
+        this.graph.centerGraph();
+    };
+
+    clearGraph = () => {
+        this.graph.clearGraph();
     };
 
     updateLayoutTimer = () => {
         this.layoutUpdateTimerID = setInterval(() => {
             if (this.vertexChanges > 0 && !this.paused) {
-                this.cy.layout(this.layout).run();
+                this.graph.updateLayout();
                 this.vertexChanges = 0;
             }
         }, 10000);
     };
 
-    start = () => {
-        this.cy = cytoscape({
-            container: document.getElementById('branchVisualizer'), // container to render in
-            style: [
-                // the stylesheet for the graph
-                {
-                    selector: 'node',
-                    style: {
-                        'background-color': '#2E8BC0',
-                        shape: 'rectangle',
-                        width: 25,
-                        height: 15
-                    }
-                },
-                {
-                    selector: 'edge',
-                    style: {
-                        width: 1,
-                        'curve-style': 'bezier',
-                        'line-color': '#696969',
-                        'control-point-step-size': '10px',
-                        events: 'no'
-                    }
-                },
-                {
-                    selector: 'node:selected',
-                    style: {
-                        'background-opacity': 0.333,
-                        'background-color': 'red'
-                    }
-                }
-            ],
-            layout: {
-                name: 'dagre'
+    trimBranchToVerticesLimit() {
+        if (this.branchOrder.length >= this.maxBranchVertices) {
+            const removeStartIndex =
+                this.branchOrder.length - this.maxBranchVertices;
+            const removed = this.branchOrder.slice(0, removeStartIndex);
+            this.branchOrder = this.branchOrder.slice(removeStartIndex);
+            this.removeBranches(removed);
+        }
+    }
+
+    removeBranches(removed: string[]) {
+        removed.forEach((id: string) => {
+            const b = this.branches.get(id);
+            if (b) {
+                this.removeVertex(id);
+                this.branches.delete(id);
             }
         });
-        this.layout = dagreOptions;
-        this.layoutApi = this.cy.layoutUtilities({
-            desiredAspectRatio: 1,
-            polyominoGridSizeFactor: 1,
-            utilityFunction: 0,
-            componentSpacing: 200
-        });
+    }
+
+    start = () => {
+        this.graph = new cytoscapeLib([dagre, layoutUtilities], initBranchDAG);
 
         // add master branch
         const master: branchVertex = {
@@ -337,7 +260,7 @@ export class BranchStore {
             '4uQeVj5tqViQh7yWWGStvkEG1Zmhx6uasJtWCJziofM',
             master
         );
-        this.cy.add({
+        this.graph.drawVertex({
             data: {
                 id: '4uQeVj5tqViQh7yWWGStvkEG1Zmhx6uasJtWCJziofM',
                 label: 'master'
@@ -345,12 +268,12 @@ export class BranchStore {
             style: {
                 'background-color': '#616161',
                 label: 'master'
-            },
-            classes: 'top-center'
+            }
         });
+        this.graph.centerVertex(master.ID);
 
         // set up click event.
-        this.cy.on('select', 'node', evt => {
+        this.graph.addNodeEventListener('select', evt => {
             const node = evt.target;
             const nodeData = node.json();
 
@@ -358,7 +281,7 @@ export class BranchStore {
         });
 
         // clear selected node.
-        this.cy.on('unselect', 'node', () => {
+        this.graph.addNodeEventListener('unselect', () => {
             this.clearSelected();
         });
