+++ plugins/dagsvisualizer/frontend/src/utils/constants.ts
@@ -1,0 +1,2 @@
+export const MAX_VERTICES = 500;
+export const DEFAULT_DASHBOARD_URL = 'http://localhost:8081';
